requirements.txt
httpx
Pillow>=11.1.0
numpy
ujson
basicsr
realesrgan